refactor(checker): split class prescan base lookup
