+++ crates/tsz-checker/src/types/class_type/core.rs
@@ -367,35 +367,8 @@ impl<'a> CheckerState<'a> {
                 }
             }
 
-            let base_prescan_type = if needs_inherited_prescan_this {
-                class
-                    .heritage_clauses
-                    .as_ref()
-                    .and_then(|heritage_clauses| {
-                        heritage_clauses.nodes.iter().find_map(|&clause_idx| {
-                            let clause_node = self.ctx.arena.get(clause_idx)?;
-                            let heritage = self.ctx.arena.get_heritage_clause(clause_node)?;
-                            if heritage.token != SyntaxKind::ExtendsKeyword as u16 {
-                                return None;
-                            }
-                            let &type_idx = heritage.types.nodes.first()?;
-                            let type_node = self.ctx.arena.get(type_idx)?;
-                            let (expr_idx, type_arguments) = if let Some(expr_type_args) =
-                                self.ctx.arena.get_expr_type_args(type_node)
-                            {
-                                (
-                                    expr_type_args.expression,
-                                    expr_type_args.type_arguments.as_ref(),
-                                )
-                            } else {
-                                (type_idx, None)
-                            };
-                            self.base_instance_type_from_expression(expr_idx, type_arguments)
-                        })
-                    })
-            } else {
-                None
-            };
+            let base_prescan_type =
+                self.inherited_prescan_this_base_type(class, needs_inherited_prescan_this);
 
             if !prescan_props.is_empty() || base_prescan_type.is_some() {
                 let own_prescan_type = if !prescan_props.is_empty() {
